Move CanonicalWeapon model into Canonical namespace
